object_detection_project/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Label Studio
label-studio>=1.10.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.11.0
psutil>=5.9.0
pynvml>=11.5.0

# Optional: For deployment
flask>=3.0.0
flask-cors>=4.0.0

